feat: can delete role

+++ app/Actions/Admin/DeleteRole.php
@@ -2,13 +2,31 @@
 
 namespace App\Actions\Admin;
 
-use App\Models\Role;
+use App\Exceptions\CannotDeleteRoleException;
+use App\Models\{Role, User};
 use App\Contracts\Actions\Admin\DeleteRoles;
 
 class DeleteRole implements DeleteRoles
 {
-    public function delete(Role $role)
+    /**
+     * @param Role $role
+     * @param User $user
+     * @return bool|null
+     * @throws CannotDeleteRoleException
+     */
+    public function delete(Role $role, User $user): ?bool
     {
+        if ($user->isNotAdmin()) {
+            throw CannotDeleteRoleException::onlyAdminCan();
+        }
+
+        $this->deleteBelongsUsers($role);
+
         return $role->delete();
+    }
+
+    public function deleteBelongsUsers(Role $role): void
+    {
+        $role->users()->detach();
     }
 }
